Add message for systems without accelerometer
Check for an accelerometer in the known location for XO-1.75 and XO-4,
and if not present give an explanation.

However, when joining a shared activity an accelerometer will be present
at the leader host, so allow the activity to start and display that
accelerometer only.

Fixes https://github.com/sugarlabs/level-activity/issues/5